# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Game On Tabletop API
  description: REST API for Game On Tabletop — a platform for organising tabletop
    game nights, tracking game collections, scheduling events, and managing group
    RSVPs.
  version: 1.0.0
paths:
...
  /api/friends:
    get:
      tags:
      - friends
      summary: List accepted friends
      operationId: list_friends_api_friends_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FriendResponse'
                type: array
                title: Response List Friends Api Friends Get
        '401':
          description: Unauthorized
      security:
      - HTTPBearer: []
  /api/friends/request:
    post:
      tags:
      - friends
      summary: Send a friend request by username
      operationId: send_friend_request_api_friends_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequestCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestResponse'
        '400':
          description: Cannot send request (user not found, self-request, or blocked)
        '401':
          description: Unauthorized
        '409':
          description: Friend request already pending
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/friends/requests:
    get:
      tags:
      - friends
      summary: List pending incoming friend requests
      operationId: list_friend_requests_api_friends_requests_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FriendResponse'
                type: array
                title: Response List Friend Requests Api Friends Requests Get
        '401':
          description: Unauthorized
      security:
      - HTTPBearer: []
  /api/friends/invite:
    post:
      tags:
      - friends
      summary: Create a shareable friend invite link
      operationId: create_friend_invite_api_friends_invite_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendInviteResponse'
        '401':
          description: Unauthorized
      security:
      - HTTPBearer: []
  /api/friends/invite/{token}:
    get:
      tags:
      - friends
      summary: Preview a friend invite link before accepting
      operationId: preview_friend_invite_api_friends_invite__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendInvitePreview'
        '404':
          description: Invite not found or expired
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/friends/invite/{token}/accept:
    post:
      tags:
      - friends
      summary: Accept a friend invite link
      operationId: accept_friend_invite_api_friends_invite__token__accept_post
      security:
      - HTTPBearer: []
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Cannot connect (self-invite or blocked)
        '401':
          description: Unauthorized
        '404':
          description: Invite not found or expired
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/friends/{friendship_id}:
    patch:
      tags:
      - friends
      summary: Accept, decline, or block a friendship
      operationId: patch_friendship_api_friends__friendship_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: friendship_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Friendship Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendshipPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not part of this friendship or wrong role
        '404':
          description: Friendship not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - friends
      summary: Remove a friendship
      operationId: delete_friendship_api_friends__friendship_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: friendship_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Friendship Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Unauthorized
        '403':
          description: Not part of this friendship
        '404':
          description: Friendship not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/friends/collection:
    get:
      tags:
      - friends
      summary: Get games owned by friends (filterable by player count)
      operationId: friends_collection_api_friends_collection_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: own
          title: Status
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: min_players
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Players
      - name: max_players
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Players
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendGameEntry'
                title: Response Friends Collection Api Friends Collection Get
        '401':
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/friends/suggestions:
    get:
      tags:
      - friends
      summary: Suggest people to add as friends based on shared groups
      operationId: friend_suggestions_api_friends_suggestions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Friend Suggestions Api Friends Suggestions Get
        '401':
          description: Unauthorized
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    FriendGameEntry:
      properties:
        bgg_game_id:
          type: integer
          title: Bgg Game Id
        game_title:
          type: string
          title: Game Title
        game_thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Game Thumbnail Url
        min_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Players
        max_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Players
        complexity:
          anyOf:
          - type: number
          - type: 'null'
          title: Complexity
        owners:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Owners
      type: object
      required:
      - bgg_game_id
      - game_title
      - game_thumbnail_url
      - min_players
      - max_players
      - complexity
      - owners
      title: FriendGameEntry
    FriendInvitePreview:
      properties:
        inviter_id:
          type: string
          title: Inviter Id
        inviter_username:
          type: string
          title: Inviter Username
        inviter_display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Inviter Display Name
        inviter_avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Inviter Avatar Url
      type: object
      required:
      - inviter_id
      - inviter_username
      - inviter_display_name
      - inviter_avatar_url
      title: FriendInvitePreview
    FriendInviteResponse:
      properties:
        token:
          type: string
          title: Token
        url:
          type: string
          title: Url
      type: object
      required:
      - token
      - url
      title: FriendInviteResponse
    FriendRequestCreate:
      properties:
        username:
          type: string
          title: Username
      type: object
      required:
      - username
      title: FriendRequestCreate
    FriendRequestResponse:
      properties:
        friendship_id:
          type: string
          title: Friendship Id
        status:
          type: string
          title: Status
      type: object
      required:
      - friendship_id
      - status
      title: FriendRequestResponse
    FriendResponse:
      properties:
        friendship_id:
          type: string
          format: uuid
          title: Friendship Id
        user_id:
          type: string
          format: uuid
          title: User Id
        username:
          type: string
          title: Username
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        status:
          type: string
          title: Status
        game_count:
          type: integer
          title: Game Count
          default: 0
      type: object
      required:
      - friendship_id
      - user_id
      - username
      - display_name
      - avatar_url
      - status
      title: FriendResponse
    FriendshipPatch:
      properties:
        status:
          type: string
          enum:
          - accepted
          - declined
          - blocked
          title: Status
      type: object
      required:
      - status
      title: FriendshipPatch
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer